requirements.txt
# Core chess functionality
python-chess[engine]>=1.999

# Rich terminal UI
rich>=13.0.0

# LLM Providers (optional - install as needed)
openai>=1.0.0
anthropic>=0.3.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Type checking (optional)
mypy>=1.0.0
types-python-chess

# Linting and formatting (optional)
black>=23.0.0
flake8>=6.0.0
isort>=5.0.0
